+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## 0.1.0 (in development)
+* Add support for OpenAI's Responses API and set it as the default.
+* Remove support for OpenAI's Chat Completions and Assistants APIs.
 
 ## 0.0.10 (2025-06-02)
 * Add new capabilities.